rev/rev-ppc64lel/solver/requirements.txt
pycryptodome~=3.22.0
